add qwen25-vl processor
