+++ fastdeploy/input/mm_processor/process_video.py
@@ -16,15 +16,74 @@
 
 import io
 import os
-import random
+import decord
+# from .utils.io_utils import EXTRACTED_FRAME_DIR, get_filename
+# from .utils.video_utils import VideoReaderWrapper
+from tempfile import NamedTemporaryFile as ntf
 
-import numpy as np
-from PIL import Image
+try:
+    # moviepy 1.0
+    import moviepy.editor as mp
+except:
+    # moviepy 2.0
+    import moviepy as mp
 
 from fastdeploy.utils import data_processor_logger
 
-from .utils.io_utils import EXTRACTED_FRAME_DIR, get_filename
-from .utils.video_utils import VideoReaderWrapper
+
+def is_gif(data: bytes) -> bool:
+    """
+    check if a bytes is a gif based on the magic head
+    """
+    return data[:6] in (b"GIF87a", b"GIF89a")
+
+
+class VideoReaderWrapper(decord.VideoReader):
+    """
+    Solving memory leak bug
+
+    https://github.com/dmlc/decord/issues/208
+    """
+
+    def __init__(self, video_path, *args, **kwargs):
+        with ntf(delete=True, suffix=".gif") as gif_file:
+            gif_input = None
+            self.original_file = None
+            if isinstance(video_path, str):
+                self.original_file = video_path
+                if video_path.lower().endswith(".gif"):
+                    gif_input = video_path
+            elif isinstance(video_path, bytes):
+                if is_gif(video_path):
+                    gif_file.write(video_path)
+                    gif_input = gif_file.name
+            elif isinstance(video_path, io.BytesIO):
+                video_path.seek(0)
+                tmp_bytes = video_path.read()
+                video_path.seek(0)
+                if is_gif(tmp_bytes):
+                    gif_file.write(tmp_bytes)
+                    gif_input = gif_file.name
+
+            if gif_input is not None:
+                clip = mp.VideoFileClip(gif_input)
+                mp4_file = ntf(delete=False, suffix=".mp4")
+                clip.write_videofile(mp4_file.name, verbose=False, logger=None)
+                clip.close()
+                video_path = mp4_file.name
+                self.original_file = video_path
+
+            super().__init__(video_path, *args, **kwargs)
+            self.seek(0)
+
+    def __getitem__(self, key):
+        frames = super().__getitem__(key)
+        self.seek(0)
+        return frames
+
+    def __del__(self):
+        if self.original_file and os.path.exists(self.original_file):
+            os.remove(self.original_file)
 
 
 def read_video_decord(video_path):
